\name{internals}
\alias{internals}

\title{Internal Functions}
%% This page is not *needed*, since all the functions are namespace-hidden
%% --  OTOH, it shows what we might want to export {and then document}
\description{
  These functions are for internal use only or not yet documented.
}
\author{Andy Leung}

\keyword{internal}

